requirements.txt
Django>=4.2.0,<5.0
Pillow>=10.0.0
python-decouple>=3.8
openpyxl>=3.1.0
boto3>=1.34.0
django-storages>=1.14.0

